net: ipv6: routing: move checking for an own source address
When IPv6 packet is received, there is a check of the packet's source
address to verify that it is not interface's non-tentative address.

This commit moves this check to the later stages of processing as
packets that can be routed are dropped in the early stage otherwise.

(cherry picked from commit c2747acbfffc564ca1545811d708f9436562a6a2)

Original-Signed-off-by: Konrad Derda <konrad.derda@nordicsemi.no>
GitOrigin-RevId: c2747acbfffc564ca1545811d708f9436562a6a2
Change-Id: I6abb95debd86520ea4c26b0d2d90fa1b7ce84beb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5745048
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>