soc/espressif/esp32c6: Do not set HAS_PM or HAS_POWEROFF
The source files required for this features are not present
in the tree for this SOC.
So if CONFIG_PM or CONFIG_POWEROFF are enabled, there would
be a cmake failure.

Let's just indicate these features are not supported in
kconfig.

(cherry picked from commit c346c77a2a0972836567a47f7656e7444878a1df)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: c346c77a2a0972836567a47f7656e7444878a1df
Change-Id: Ia1d2160088ce54345f13413fc8588fc990b01cfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5716813
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>