arch: riscv: stacktrace: fix output without `ra` on the stack top
Account for the scenario when we are doing `esf`-based
unwinding from a function which doesn't have any callee.
In this case the `ra` is not saved on the stack and the
second function from the top of the frame could be missing.

(cherry picked from commit 9698df9dc42b43711f7644217408f6a9cddfe37b)

Original-Signed-off-by: Yong Cong Sin <ycsin@meta.com>
Original-Signed-off-by: Yong Cong Sin <yongcong.sin@gmail.com>
GitOrigin-RevId: 9698df9dc42b43711f7644217408f6a9cddfe37b
Cr-Build-Id: 8740269368881080913
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8740269368881080913
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8a05233ce5b3d41e3839a7d456b2bf5c4579f866
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5769424
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>