Bluetooth: TBS: Add support for dynamic TBS
The TBS and GTBS instances can now be dynamically registered
and unregistered, which fits better for the use cases of TBS
where specific bearers can come and go depending on the
features of the device.

For example if a SIM card is inserted, then a device can
register a TBS for the provider and remove it again if the
SIM card is removed.

(cherry picked from commit 1406c547d63417d421ba76ca1ccd1de12e32126c)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 1406c547d63417d421ba76ca1ccd1de12e32126c
Cr-Build-Id: 8735895478936220081
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735895478936220081
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic2a326908c45cd6da6fc5894d4dd71f8320a3e40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5887642
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>