net: coap_client: handle truncation of received data
Use the MSG_TRUNC flag to check the total length of UDP packets
receieved by the coap_client, and hence check if the receive buffer
contains the whole message, or if it is truncated.

If the message is truncated, then use a blockwise transfer to fetch the
rest of the data.

This is related to issue #76089.

(cherry picked from commit 93f871e2d167578a5fcaafc510ac3edff609710f)

Original-Signed-off-by: Matt Rodgers <mrodgers@witekio.com>
GitOrigin-RevId: 93f871e2d167578a5fcaafc510ac3edff609710f
Change-Id: Ib8450e15e7c6814bb5e1b67de447d0d48bca7140
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5752453
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>