net: coap_client: limit payload size during block transfer to block size
Limit the coap payload size passed up to the application callback to the
block size, when a block transfer is in progress and the current payload
is not the final block.

If the current payload is not part of a block transfer, or is the final
block of a transfer, then the full payload can be passed to the
application to avoid having to make another request over the network for
data that has already been received.

This avoids a problem raised in issue #76089, where a payload longer
than CONFIG_COAP_CLIENT_MESSAGE_SIZE causes the same data to be passed
to the application callback twice (once in the large packet, and once in
the next block which must have an offset that is a multiple of the block
size).

(cherry picked from commit 44d39e20284ecdc3f93c872a1f144b48ea13ae78)

Original-Signed-off-by: Matt Rodgers <mrodgers@witekio.com>
GitOrigin-RevId: 44d39e20284ecdc3f93c872a1f144b48ea13ae78
Change-Id: Id94586095ee94e1511733137a755c4538b578abe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5752454
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>