drivers: gnss: api: Remove periodic configuration
The GNSS device driver API includes an API for setting the periodic
power configuration at runtime. This API is not used in-tree, and is
antithetical to the pm device subsystem.

The periodic config configures the GNSS to search for a fix for a
given amount of time, then suspend itself for a given amount of time,
and repeat this pattern. The issue with this is that both the
application and the GNSS modem itself can be in a conflicting power
state, where the device is resumed (pm device), but actually suspended.

Applications which wish to periodically resume/suspend the modem, even
in more sophisticated ways like suspended after first fix, can simply
implement this logic, using the pm_device calls and the GNSS callbacks.

(cherry picked from commit a28ba1d1a27e8c70359b4609b67fc74b07e0e7f1)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
GitOrigin-RevId: a28ba1d1a27e8c70359b4609b67fc74b07e0e7f1
Cr-Build-Id: 8737007880308062673
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737007880308062673
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I623dd0ca1c31f9289a0dbfe404c443846c3ec741
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5845115
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>