drivers/interrupt: it8xxx2: Register interrupt number 0 to handle
In the it8xxx2 chip, the interrupt for INT0 is reserved. However,in some
stress tests, the unhandled IRQ0 issue occurs. To prevent the system from
going directly into kernel panic, we implemented a workaround by
registering interrupt number 0 and doing nothing in the IRQ0 handler.
The side effect of this solution is that when IRQ0 is triggered, it will
take some time to execute the routine. There is no need to worry about
missing interrupts because each IRQ's ISR is write-clear, and if the
status is not cleared, it will continue to trigger.

(cherry picked from commit 0980cc4390eaf163f611012faf5a2cc3fd046942)

Original-Signed-off-by: Tim Lin <tim2.lin@ite.corp-partner.google.com>
GitOrigin-RevId: 0980cc4390eaf163f611012faf5a2cc3fd046942
Cr-Build-Id: 8740812952977153137
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8740812952977153137
Change-Id: I0e43401ab4e43a63e1b368e3c042ef3e8c7e168c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5756308
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>