drivers: wifi: eswifi: Fix memory buffer allocation in off_read_work
When receiving data over the eswifi module, we currently read the data
first, then allocate a buffer, and finally write the data into the
buffer. The issue is that if we can't allocate the buffer, the data
that was read is lost. To fix this, we should first attempt to allocate
the buffer before reading any data. If we can't allocate the buffer, we
should not proceed with reading the data. By allocating a buffer with
the MTU size, we can read the packet, write it into the allocated buffer
and then resize by removing unused allocated buffer with
net_pkt_trim_buffer().

(cherry picked from commit cf0c2a527593fa1a90d13dd08f1fbbe08a193c53)

Original-Signed-off-by: Léo BRIAND <leo.briand@smile.fr>
GitOrigin-RevId: cf0c2a527593fa1a90d13dd08f1fbbe08a193c53
Cr-Build-Id: 8740450565536499009
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8740450565536499009
Change-Id: I49ec0bd0e741c060658d0d9f40c7d69234aef01b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5761297
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>