Bluetooth: Audio: Rename bt_audio_codec_qos -> bt_bap_qos_cfg
The QoS structure is not related to a codec, but rather a
stream, and should thus not use the "Codec" name.

The BAP and ASCS specs refer to the QoS as
"QoS configuration" several places, so it is an obvious
choice for a name.

Since the structure is defined and used by BAP, the prefix
was changed from bt_audio to bt_bap.

(cherry picked from commit dc97bbd35f6bca23aed1364550bb9b2c74abd904)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: dc97bbd35f6bca23aed1364550bb9b2c74abd904
Cr-Build-Id: 8734742957433546641
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8734742957433546641
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iea2c1708317e99068637541a79665e483a32fdc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5908562
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>