drivers: usb device: Remove MCUX controller type
The CONFIG_USB_MCUX_CONTROLLER_TYPE choice config
is not necessary, let's just remove it. Theoretically,
if there was an SOC with the EHCI and LPCIP3511, then
it might cause a build error, but there is no SOC
which is supported on this legacy driver that has that.

Remove all settings of it in the SOC files.

(cherry picked from commit 4e9227641ee3585552619322049c75f8f5d7f076)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 4e9227641ee3585552619322049c75f8f5d7f076
Cr-Build-Id: 8738910416953667585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8738910416953667585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I23ac3810e6627b74a2ad72b5e5079de1944fa32b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5806804
Commit-Queue: Eric Yilun Lin <yllin@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Eric Yilun Lin <yllin@google.com>