adc: current_sense_amplifier: resistance in milli-ohms
Change the unit of the sense resistor in the devicetree binding from
micro-ohms to milli-ohms. This is done for three reasons.

Firstly, the maximum value resistor that can currently be represented
is 4.2 kOhms, due to the limitation of devicetree properties to 32 bits.

Secondly, storing the resistance at such a high resolution makes
overflows much more likely when the desired output unit is micro-amps,
not milli-amps.

Finally, micro-ohms, are an unnecessarily precise unit for the purpose
of these calculations, and a resolution that is not realistic to
achieve. The high resistor resolution results in large divisors that
reduce the resolution of outputs. Unlike resistors characterised down to
the micro-ohm, devices wanting to measure micro-amps are actually
realistic.

(cherry picked from commit 4361c96c48f30f1a39f59b268b713009a92e6d1b)

Original-Signed-off-by: Jordan Yates <jordan@embeint.com>
GitOrigin-RevId: 4361c96c48f30f1a39f59b268b713009a92e6d1b
Cr-Build-Id: 8733746390213692641
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8733746390213692641
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I40fbf570ba53f8b63480ee43b83078afa94d2b61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5934655
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>