charger: bq25180: Add threshold voltage for precharge
Charging a battery has three phases (compare Figure 8-2 in
https://www.ti.com/lit/ds/symlink/bq25180.pdf)

First is a pre-charge phase, then a constant current, then a constant
voltage phase.
During the pre-charge phase, the battery is only charged with a very low
current, to not damage the cells, because they are below a threshold,
that is considered healthy for the battery and need to be brought back
up to a healthy voltage.

(cherry picked from commit cdfab15b999b940e9457ca5167fbae02a47915a9)

Original-Signed-off-by: Fabian Pflug <fabian.pflug@grandcentrix.net>
GitOrigin-RevId: cdfab15b999b940e9457ca5167fbae02a47915a9
Cr-Build-Id: 8737046051874902001
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737046051874902001
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I403c4f3e21a03cf0634b2ee3cc306d8b7179d4d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5842124
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>