Bluetooth: doc: update kconfig docstring
CONFIG_BT_CONN_TX_USER_DATA_SIZE is now used for callbacks in the host.
We don't want to limit ourselves to that, so change the wording to be
more generic.
In the future, the plan is to not use user_data at all, removing the
need for the kconfig altogether.

Also rename the structure that was referenced in the docstring.

(cherry picked from commit 6b6c4333b7fb85d7ea0729e2d084af94f3b65ac4)

Original-Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>
GitOrigin-RevId: 6b6c4333b7fb85d7ea0729e2d084af94f3b65ac4
Cr-Build-Id: 8739001010820476625
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739001010820476625
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie127f8000166d57d345355a7e03d0a2e5be82ac8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5803467
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Commit-Queue: Eric Yilun Lin <yllin@google.com>