drivers: led: lp5562: reuse engine already linked
Right now "lp5562_led_blink" unconditionally find and assign a engine
that is not executing at the moment. This is an issue, if 3 "led" are
already using an engine. That is a call to "lp5562_led_blink" could
terminate with error, even if the "led" is already using an engine.
This commit fixes this issue by adding a check to see if engine is
linked and reuse it, otherwise assign a free engine.

(cherry picked from commit 645beff897eb0babb50fa865808e2ba547147766)

Original-Signed-off-by: Anuj Pathak <anuj@croxel.com>
GitOrigin-RevId: 645beff897eb0babb50fa865808e2ba547147766
Cr-Build-Id: 8737098476786990081
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737098476786990081
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic77383a944231a20c6d4b4ee2880312a459922fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5839800
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>