llext: add bringup, teardown, and bootstrap APIs
llext_bringup() and llext_teardown() are intended to be used to call the
extension's own initialization and cleanup functions, respectively. They
are meant to be called by the developer after loading an extension and
before unloading it. The list of function pointers to be called is
obtained via the new llext_get_fn_table() syscall, so that they are
compatible with user mode.

llext_bootstrap() is intended to be used as the entry point for a thread
created to run an extension, in either user or kernel contexts. It will
call the extension's own initialization functions and then an additional
entry point in the same context (if desired). The same function can also
be called directly in the main thread, if only initialization is
required.

(cherry picked from commit af302cd5fe6b8bad2251a23362f1015c3c44309e)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: af302cd5fe6b8bad2251a23362f1015c3c44309e
Cr-Build-Id: 8737046051874902001
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737046051874902001
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I88c5df27ce270941294702b4a11fbbb7713ea180
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5843376
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>