Bluetooth: Mesh: Move testing.h to mesh subsys instead of include
The file is only ever used by mesh and it is even stated in the
header of the file that it is internal APIs.

The include file has been merged with the host testing.h file
as that was mesh specific as well. Similarly the testing.c
file was also moved

This is part of a process to clean up the file structure of
Bluetooth as it's a bit messy, which is evident from the
MAINTAINERS.yml file.

(cherry picked from commit f1b182b5856434c55fb688919bf163d274e8a8ac)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: f1b182b5856434c55fb688919bf163d274e8a8ac
Cr-Build-Id: 8738366835818535617
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8738366835818535617
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6ac15f168b37be8ac5ae9f6a1c5b1ce9ccd88b88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5822414
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>