doc: samples: use :zephyr-app: for in-tree samples
The zephyr-app-commands directive can output a helpful hint to the user
when they are trying to build a sample that is in the Zephyr tree,
telling them to ensure they are in the root folder of the Zephyr repo.

Update all doc pages that were using :app: instead of :zephyr-app: so
that the hint is displayed.

(cherry picked from commit 34982bb00b57486aeb83a6cb3bb7433791302aed)

Original-Signed-off-by: Benjamin Cabé <benjamin@zephyrproject.org>
GitOrigin-RevId: 34982bb00b57486aeb83a6cb3bb7433791302aed
Cr-Build-Id: 8737046051874902001
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737046051874902001
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic0edd9fed42ced743ba85688415c7859b4f0a24a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5842145
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>