tests: net: socket_tls: test different TLS configurations with sockets
This commit adds a test to showcase how to configure TLS 1.2 and 1.3
socket connections against an OpenSSL server. Only a limited number
of combinations are available for now:

- TLS 1.2
	- RSA certificate and key exchange.
	- ECDSA certificate and ECDHE key exchange.

- TLS 1.3 only supports ephemeral (ECDHE) key exchange with/without
  session tickets.

Since the goal is to test TLS connection and not low level ethernet
functionalities or similar, the only supported platform is "native_sim"
where Linux sockets are used to connect to the OpenSSL server locally.

The idea is that the Zephyr application acts a client and tries to connect
to the OpenSSL server running on the same PC. For sake of simplificity a
bash script is provided to start the OpenSSL server properly.
For completeness a bash script is also provided to re-generate
certificates and keys.

(cherry picked from commit 91fa8a2b639931f25351710cfbfe35a45c17ca4f)

Original-Signed-off-by: Valerio Setti <vsetti@baylibre.com>
GitOrigin-RevId: 91fa8a2b639931f25351710cfbfe35a45c17ca4f
Cr-Build-Id: 8736554896755799505
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8736554896755799505
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I145272ac62b2eb32d6657a22f7a648cec3febcb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5860991
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>