samples/llext/modules: build hello_world as built-in by default
This commit changes the default value of CONFIG_HELLO_WORLD_MODE to 'y'
so that the hello_world function is compiled as a built-in part of the
Zephyr image by default. This is the simplest possible configuration,
that works for all architectures.

To build hello_world as an llext module, the user must either follow the
commands in the documentation or set CONFIG_HELLO_WORLD_MODE=m in the
project file along with additional architecture-specific settings that
may be required for proper LLEXT support on the target.
Leave a note in the prj.conf and sample.yaml files to remind the user.

(cherry picked from commit 030c1c83082a92ad1f5775e8dcfbafc431bacf00)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: 030c1c83082a92ad1f5775e8dcfbafc431bacf00
Cr-Build-Id: 8737823252869148017
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737823252869148017
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If10b75b7aecb0517512b4abdeeacfcc375556e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5829311
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>