Bluetooth: BAP: Broadcast sync fail receive state bis sync value fixed
In the case that a broadcast sync fails for any reason, we need to
set the BIS_Sync value for all subgroups to 0xFFFFFFFF
(BT_BAP_BIS_SYNC_FAILED) as per the BAP spec.

This commit adds a new define for this and modifies
both the scan delegator and broadcast sink to support this.

The change is validated by a modification to the broadcast
assistant test that verifies that the value is set in the
case of an invalid broadcast code (which indicates a
failed BIG sync).

(cherry picked from commit b800be343063a5b8c5cfe5afb071b318af36d8a1)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: b800be343063a5b8c5cfe5afb071b318af36d8a1
Cr-Build-Id: 8735998619919803697
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735998619919803697
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic7efb06dd830cf1e8691dae3161d0952be207418
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5881542
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>