tests: Bluetooth: Run Audio BSIM test on nrf5340bsim
Add support for running the audio BSIM tests on the
nrf5340bsim board.

This expands the tests we run in CI.

To support all tests with the hci_ipc sample for the netcore,
it is updated to support a bit more while still fitting on the
netcore on the nRF5340 series.

The behavior of advertisement seems to differ a little bit,
which required the bap_unicast_server_test.c to update how it
restarts advertisements.

Two broadcast tests that are passing on the nRF52 BSIM is also
failing due an assert. This should be investigated at a later
time.

(cherry picked from commit fa0bbad205ebb57ecfbd0cba17730034d18821b1)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: fa0bbad205ebb57ecfbd0cba17730034d18821b1
Cr-Build-Id: 8736917283475341809
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8736917283475341809
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I948ebf92016251a8a1de2bfd899a525b9d97ef91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5850796
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>