doc: Check redirect entries are valid
While it's by definition not possible to check that the "source" for a
redirect entry is a valid path, the "destination" could be checked to be
pointing to an existing resource. This will also help catch potential
errors in the "source" as well since it's common to mess up the paths
for both.

Fixes #77797

(cherry picked from commit 08688b1876acf2e33fdd95ec62564abadc3c8954)

Original-Signed-off-by: Benjamin Cabé <benjamin@zephyrproject.org>
GitOrigin-RevId: 08688b1876acf2e33fdd95ec62564abadc3c8954
Cr-Build-Id: 8737732657120454561
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737732657120454561
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ibfc1af01bd68f62d3355460ec3027b055d95693a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5836423
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>