wifi_mgmt: Add support for configuring PS exit strategy
If AP indicates the presence of buffered traffic, then it is up to the
STA to decide whether to stay in PS or come out of PS, add configuration
options that can be used at runtime to choose this.

This is tagged as "noup" because it's a backport and "fromlist" cannot
be used as it won't apply cleanly.

(cherry picked from commit 0ce5da6da8124dfe760e17e87e60746798aae2a3)

Original-Signed-off-by: Ajay Parida <ajay.parida@nordicsemi.no>
GitOrigin-RevId: 0ce5da6da8124dfe760e17e87e60746798aae2a3
Cr-Build-Id: 8733565196572304993
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8733565196572304993
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0bf757581f121691975ef2cfa1be38cca12e9118
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5940632
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Eric Yilun Lin <yllin@google.com>