bluetooth: mesh: brg_cfg_cli: copy buf in synchronous api
When a synchronous API is used, the content of `buf` will not be
valid by the time the thread that called the synchronous API is woken up
again.

Therefore, the simplest way to solve this is when a user allocates the
buffer which will be filled up with the content of the buffer passed to
the model callback.

(cherry picked from commit 76f015b61f69007bb28a2a99cf5bb2d5e96e4e2a)

Original-Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>
GitOrigin-RevId: 76f015b61f69007bb28a2a99cf5bb2d5e96e4e2a
Cr-Build-Id: 8736554896755799505
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8736554896755799505
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1b074e8b43488e4741e8d63e1b0f0a924e755bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5870341
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>