Bluetooth: MPL: Replace busy bool with atomic
Replace the busy boolean flag with an atomic value.
This prevents any race conditions with the MPL implementation.
Modifies where the new atomic value is set and cleared
so that initialization gets to finish before allowing
any reads.

Due to how the MPL is structured, and how a select cannot
be rejected from OTS, this does not give a perfect solution.
Ideally we need a separate object per OTS object, rather than
a shared one, and/or the OTS implemenation would allow
us to reject a select if the object is not currently
available or ready.

This commit does not fix the above issues, as that is a
larger undertaking.

(cherry picked from commit 4b68043ebbe66af6d943a908d98b25fbe4aec0b9)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 4b68043ebbe66af6d943a908d98b25fbe4aec0b9
Cr-Build-Id: 8736554896755799505
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8736554896755799505
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie615ffeffdc6fa383276f92a31ed4b812d784480
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5870279
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>