usb: device: avoid starving other threads in CDC ACM UART fifo_fill
The check for the device being configured or suspended in the fifo_fill
implementation can starve other threads, because the early return does
not change the state of the TX path, and fifo_fill is called again
from the callback loop. The ringbuffer that emulates the TX FIFO can be
filled and marked ready as long as the space is available.

(cherry picked from commit 0f0322127df5579cd3663f87039207dc7595dad1)

Original-Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>
GitOrigin-RevId: 0f0322127df5579cd3663f87039207dc7595dad1
Cr-Build-Id: 8735105345130440289
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735105345130440289
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I493919255c41a424a3ce75d383621c8b6ac9df95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5906810
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>