dts: renesas_ra: Rename dts node path
Changes the path name of a DTS node so that it can be used
as the stem of a BSP macro.
All nodes to be changed are referenced via labels,
so only the name is changed.

(cherry picked from commit 91d8e72dd47e79eee6c80dc7979df5cdd0b7c9d6)

Original-Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>
GitOrigin-RevId: 91d8e72dd47e79eee6c80dc7979df5cdd0b7c9d6
Cr-Build-Id: 8734380568354140561
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8734380568354140561
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I74454c846176cb4c71409debf8fd8a7bcf85ba73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5915892
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>